Conflicts:
	src/main/webapp/paginas/cargo/consultarCargo.xhtml

+++ src/main/java/com/indra/sishe/entity/Periodo.java
@@ -17,61 +17,75 @@ public class Periodo implements Serializable, Comparable<Periodo> {
 
 	private Date horaFim;
 
-	private Double porcentagem;
+	private Integer porcentagem;
 
 	public Periodo() {
+
 	}
 
 	public Long getId() {
+
 		return id;
 	}
 
 	public void setId(Long id) {
+
 		this.id = id;
 	}
 
 	public Regra getRegra() {
+
 		return regra;
 	}
 
 	public void setRegra(Regra regra) {
+
 		this.regra = regra;
 	}
 
 	public Integer getDiaSemana() {
+
 		return diaSemana;
 	}
 
 	public void setDiaSemana(Integer diaSemana) {
+
 		this.diaSemana = diaSemana;
 	}
 
 	public Date getHoraInicio() {
+
 		return horaInicio;
 	}
 
 	public void setHoraInicio(Date horaInicio) {
+
 		this.horaInicio = horaInicio;
 	}
 
 	public Date getHoraFim() {
+
 		return horaFim;
 	}
 
 	public void setHoraFim(Date horaFim) {
+
 		this.horaFim = horaFim;
 	}
 
-	public Double getPorcentagem() {
+	public Integer getPorcentagem() {
+
 		return porcentagem;
 	}
 
-	public void setPorcentagem(Double porcentagem) {
+	public void setPorcentagem(Integer porcentagem) {
+
 		this.porcentagem = porcentagem;
 	}
 
 	@Override
 	public int hashCode() {
+
 		final int prime = 31;
 		int result = 1;
 		result = prime * result + ((diaSemana == null) ? 0 : diaSemana.hashCode());
@@ -83,38 +97,29 @@ public class Periodo implements Serializable, Comparable<Periodo> {
 
 	@Override
 	public boolean equals(Object obj) {
-		if (this == obj)
-			return true;
-		if (obj == null)
-			return false;
-		if (getClass() != obj.getClass())
-			return false;
+
+		if (this == obj) return true;
+		if (obj == null) return false;
+		if (getClass() != obj.getClass()) return false;
 		Periodo other = (Periodo) obj;
 		if (diaSemana == null) {
-			if (other.diaSemana != null)
-				return false;
-		} else if (!diaSemana.equals(other.diaSemana))
-			return false;
+			if (other.diaSemana != null) return false;
+		} else if (!diaSemana.equals(other.diaSemana)) return false;
 		if (horaFim == null) {
-			if (other.horaFim != null)
-				return false;
-		} else if (!horaFim.equals(other.horaFim))
-			return false;
+			if (other.horaFim != null) return false;
+		} else if (!horaFim.equals(other.horaFim)) return false;
 		if (horaInicio == null) {
-			if (other.horaInicio != null)
-				return false;
-		} else if (!horaInicio.equals(other.horaInicio))
-			return false;
+			if (other.horaInicio != null) return false;
+		} else if (!horaInicio.equals(other.horaInicio)) return false;
 		if (id == null) {
-			if (other.id != null)
-				return false;
-		} else if (!id.equals(other.id))
-			return false;
+			if (other.id != null) return false;
+		} else if (!id.equals(other.id)) return false;
 		return true;
 	}
 
 	@Override
 	public int compareTo(Periodo o) {
+
 		int valor = diaSemana.compareTo(o.diaSemana);
 		if (valor != 0) {
 			return valor;
